% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_main_function.R
\name{quantile.subsampling}
\alias{quantile.subsampling}
\title{Optimal Subsampling Methods for Quantile Models}
\usage{
\method{quantile}{subsampling}(
  formula,
  data,
  tau,
  n.plt,
  n.ssp,
  B,
  criterion = c("OptL"),
  sampling.method = c("WithReplacement"),
  estimate.method = c("Weighted", "Uniform"),
  parallel = parallel
)
}
\arguments{
\item{formula}{An object of class "formula" which describes the model to be
fitted.}

\item{data}{A data frame containing the variables in the model. Usually it
contains a response vector and a design matrix.
The binary response vector that takes the value of 0 or 1, where 1 means
the event occurred.
The design matrix contains predictor variables. A column representing
the intercept term with all 1's will be automatically added.}

\item{tau}{The quantile.}

\item{n.plt}{The pilot subsample size (the first-step subsample size).
These samples will be used to estimate the pilot estimator.}

\item{n.ssp}{The expected optimal subsample size (the second-step subsample
size).}

\item{B}{TBD}

\item{criterion}{The criterion of optimal subsampling probabilities.}

\item{sampling.method}{The sampling method for drawing the optimal subsample.}

\item{estimate.method}{The type of the maximum likelihood function used to
calculate the optimal subsampling estimator.}
}
\value{
\describe{
\item{model.call}{the model}
\item{beta.plt}{pilot estimator}
\item{beta.ssp}{optimal subsample estimator}
\item{est.cov.ssp}{covariance matrix of \code{beta.ssp}}
\item{index.plt}{index of pilot subsample}
\item{index.ssp}{index of optimal subsample}
}
}
\description{
Optimal Subsampling Methods for Quantile Models
}
\details{
Additional details... briefly introduce the idea.
}
\examples{
#quantile regression
set.seed(1)
N <- 1e6
B <- 10
tau <- 0.5
beta.true <- rep(1, 7)
d <- length(beta.true) - 1
corr  <- 0.5
sigmax  <- matrix(0, d, d)
for (i in 1:d) for (j in 1:d) sigmax[i, j] <- corr^(abs(i-j))
X <- MASS::mvrnorm(N, rep(0, d), sigmax)
err <- rnorm(N, 0, 1) - qnorm(tau)
Y <- beta.true[1] + X \%*\% beta.true[-1] + 
err * rowMeans(abs(X))
data <- as.data.frame(cbind(Y, X))
formula <- Y ~ X
n.plt <- 1000
n.ssp <- 1000
optL.results <- quantile.subsampling(formula,data,tau = tau,n.plt = n.plt,
n.ssp = n.ssp,B,criterion = 'OptL',sampling.method = 'WithReplacement',
estimate.method = 'Weighted',parallel = F)
uni.results <- quantile.subsampling(formula,data,tau = tau,n.plt = n.plt,
n.ssp = n.ssp,B,sampling.method = 'WithReplacement',
estimate.method = 'Uniform',parallel = F)
}
