% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/softmax_functions.R
\name{softmax.summary}
\alias{softmax.summary}
\title{Softmax Main results summary}
\usage{
softmax.summary(object)
}
\arguments{
\item{object}{A list object output by the main function, which contains the
results of the estimation of the parameters, the estimation of the
variance, subsample size, etc.}
}
\value{
A series of data.frame will be printed.
}
\description{
Softmax Main results summary
}
\examples{
#TBD
}
