% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_main_function.R
\name{ssp.quantreg}
\alias{ssp.quantreg}
\title{Optimal Subsampling Methods for Quantile Models}
\usage{
ssp.quantreg(
  formula,
  data,
  subset = NULL,
  tau = 0.5,
  n.plt,
  n.ssp,
  B = 5,
  boot = TRUE,
  criterion = "optL",
  sampling.method = "withReplacement",
  likelihood = c("weighted"),
  control = list(...),
  contrasts = NULL,
  ...
)
}
\arguments{
\item{formula}{An object of class "formula" which describes the model to be
fitted.}

\item{data}{A data frame containing the variables in the model. Usually it
contains a response vector and a design matrix.
The binary response vector that takes the value of 0 or 1, where 1 means
the event occurred.
The design matrix contains predictor variables. A column representing
the intercept term with all 1's will be automatically added.}

\item{subset}{An optional vector specifying a subset of rows to be used}

\item{tau}{The quantile.}

\item{n.plt}{The pilot subsample size (the first-step subsample size).
These samples will be used to estimate the pilot estimator.}

\item{n.ssp}{The expected optimal subsample size (the second-step subsample
size).}

\item{B}{TBD}

\item{boot}{TBD}

\item{criterion}{The criterion of optimal subsampling probabilities.}

\item{sampling.method}{The sampling method for drawing the optimal subsample.}

\item{likelihood}{The type of the maximum likelihood function used to
calculate the optimal subsampling estimator.}

\item{control}{a list of parameters for controlling the fitting process.}

\item{contrasts}{The type of the maximum likelihood function used to}

\item{...}{a list of parameters for controlling the fitting process.}
}
\value{
\describe{
\item{model.call}{the model}
\item{beta.plt}{pilot estimator}
\item{beta.ssp}{optimal subsample estimator}
\item{est.cov.ssp}{covariance matrix of \code{beta.ssp}}
\item{index.plt}{index of pilot subsample}
\item{index.ssp}{index of optimal subsample}
}
}
\description{
Optimal Subsampling Methods for Quantile Models
}
\details{
Additional details... briefly introduce the idea.
}
\examples{
#quantile regression
set.seed(1)
N <- 1e4
B <- 5
tau <- 0.75
beta.true <- rep(1, 7)
d <- length(beta.true) - 1
corr  <- 0.5
sigmax  <- matrix(0, d, d)
for (i in 1:d) for (j in 1:d) sigmax[i, j] <- corr^(abs(i-j))
X <- MASS::mvrnorm(N, rep(0, d), sigmax)
err <- rnorm(N, 0, 1) - qnorm(tau)
Y <- beta.true[1] + X \%*\% beta.true[-1] + 
err * rowMeans(abs(X))
data <- as.data.frame(cbind(Y, X))
formula <- Y ~ X
n.plt <- 200
n.ssp <- 100
optL.results <- ssp.quantreg(formula,data,tau = tau,n.plt = n.plt,
n.ssp = n.ssp,B = B,boot = TRUE,criterion = 'optL',
sampling.method = 'withReplacement',likelihood = 'weighted')
summary(optL.results)
uni.results <- ssp.quantreg(formula,data,tau = tau,n.plt = n.plt,
n.ssp = n.ssp,B = B,boot = TRUE,criterion = 'uniform',
sampling.method = 'withReplacement', likelihood = 'weighted')
summary(uni.results)
}
