% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rare_events_logistic_main_function.R
\name{ssp.relogit}
\alias{ssp.relogit}
\title{Optimal Subsampling for Logistic Regression Model with Rare Events Data}
\usage{
ssp.relogit(
  formula,
  data,
  subset = NULL,
  n.plt,
  n.ssp,
  criterion = "optL",
  likelihood = "logOddsCorrection",
  alpha = 0.1,
  b = 2,
  control = list(...),
  contrasts = NULL,
  ...
)
}
\arguments{
\item{formula}{An object of class "formula" which describes the model to be
fitted.}

\item{data}{A data frame containing the variables in the model. Usually it
contains a response vector and a design matrix. The binary response vector
that takes the value of 0 or 1, where 1 means the event occurred. The
design matrix contains predictor variables. A column representing the
intercept term with all 1's will be automatically added.}

\item{subset}{An optional vector specifying a subset of rows to be used}

\item{n.plt}{The pilot subsample size (the first-step subsample size).}

\item{n.ssp}{The expected optimal subsample size (the second-step subsample
size) drawn from those samples with \code{Y=0}.}

\item{criterion}{The criterion of optimal subsampling probabilities,
currently there are three choices \code{optA}, \code{optL}, and \code{LCC}.}

\item{likelihood}{The type of the maximum likelihood function used to
calculate the optimal subsampling estimator, currently there are two choices
\code{weighted} and \code{logOddsCorrection}.}

\item{alpha}{Mixture proportions of optimal subsampling probability and
uniform subsampling probability. Default = 0.1.}

\item{b}{This parameter controls the upper threshold for optimal subsampling
probabilities.}

\item{control}{a list of parameters for controlling the fitting process.}

\item{contrasts}{The type of the maximum likelihood function used to}

\item{...}{a list of parameters for controlling the fitting process.}
}
\value{
\describe{
\item{beta.plt}{pilot estimator}
\item{beta.ssp}{optimal subsample estimator}
\item{beta.cmb}{combined estimator of \code{beta.plt} and \code{beta.ssp}}
\item{var.ssp}{covariance matrix of \code{beta.ssp}}
\item{var.cmb}{covariance matrix of \code{beta.cmb}}
\item{index.plt}{index of pilot subsample}
\item{index.ssp}{index of optimal subsample. The expectation of
\code{length(index.ssp)} is n.ssp plus the number of rare event data.}
}
}
\description{
Optimal Subsampling for Logistic Regression Model with Rare Events Data
}
\details{
Additional details... briefly introduce the idea of this method.
}
\examples{
set.seed(1)
N <- 2 * 1e4
beta0 <- c(-5, -rep(0.7, 6))
d <- length(beta0) - 1
X <- matrix(0, N, d)
corr <- 0.5
sigmax <- corr ^ abs(outer(1:d, 1:d, "-"))
sigmax <- sigmax / 4
X <- MASS::mvrnorm(n = N, mu = rep(0, d), Sigma = sigmax)
Y <- rbinom(N, 1, 1 - 1 / (1 + exp(beta0[1] + X \%*\% beta0[-1])))
print(paste('N: ', N))
print(paste('sum(Y): ', sum(Y)))
n.plt <- 200
n.ssp <- 1000
data <- as.data.frame(cbind(Y, X))
formula <- Y ~ .
subsampling.results <- ssp.relogit(formula = formula,
                                     data = data,
                                     n.plt = n.plt,
                                     n.ssp = n.ssp,
                                     criterion = 'optA',
                                     likelihood = 'logOddsCorrection')
summary(subsampling.results)
}
