% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_glm_functions.R
\name{subsampling.summary}
\alias{subsampling.summary}
\title{Main results summary}
\usage{
subsampling.summary(object)
}
\arguments{
\item{object}{A list object output by the main function, which contains the
results of the estimation of the parameters, the estimation of the
covariance matrix, subsample size, etc.}
}
\value{
A data.frame will be printed.
}
\description{
Main results summary
}
\examples{
#logistic regression
set.seed(1)
N <- 1e4
beta0 <- rep(-0.5, 7)
d <- length(beta0) - 1
X <- matrix(0, N, d)
generate_rexp <- function(x) x <- rexp(N, rate = 2)
X <- apply(X, 2, generate_rexp)
Y <- rbinom(N, 1, 1 - 1 / (1 + exp(beta0[1] + X \%*\% beta0[-1])))
print(paste('N: ', N))
print(paste('sum(Y): ', sum(Y)))
data <- as.data.frame(cbind(Y, X))
formula <- Y ~ .
n.plt <- 200
n.ssp <- 600
subsampling.results <- glm.subsampling(formula, data, n.plt, n.ssp,
family = 'binomial', criterion = "OptL", sampling.method = 'Poisson',
likelihood = "LogOddsCorrection")
subsampling.summary(subsampling.results)
}
