% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsampling_cox.R
\name{subsampling_cox}
\alias{subsampling_cox}
\title{Subsampling by Cox's method (update with Nir)}
\usage{
subsampling_cox(V, D, X, R = NULL, q, q0 = q, method)
}
\arguments{
\item{V}{observed times.}

\item{D}{status (T/F).}

\item{X}{covariate matrix.}

\item{R}{recruitment (left truncation) times.}

\item{q}{number of subsampled censored observations in the second round for the L and A.}

\item{q0}{number of subsampled censored observations for the uniform estimator (being the pilot estimator for L and A). Defaults to be the same as q..}

\item{method}{one of "U"/"L"/"A" (standing for uniform, L-optimal, A-optimal).}
}
\value{
The calculated subsampling.
}
\description{
This function *** (update with Nir).
}
\examples{
subsampling_cox(V, D, X, R, q, q0, "U")
}
