% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_functions.R
\name{summary.quantile.subsampling}
\alias{summary.quantile.subsampling}
\title{Main results summary}
\usage{
\method{summary}{quantile.subsampling}(object)
}
\arguments{
\item{object}{A list object output by the main function, which contains the
results of the estimation of the parameters, the estimation of the
covariance matrix, subsample size, etc.}
}
\value{
A data frame will be printed.
}
\description{
Main results summary
}
\examples{
#logistic regression
set.seed(1)
}
