% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_glm_functions.R
\name{summary.ssp.glm}
\alias{summary.ssp.glm}
\title{Main results summary}
\usage{
\method{summary}{ssp.glm}(object, ...)
}
\arguments{
\item{object}{A list object output by the main function, which contains the
results of the estimation of the parameters, the estimation of the
covariance matrix, subsample size, etc.}

\item{...}{Additional arguments passed to the summary function.}
}
\value{
A data.frame will be printed.
}
\description{
Main results summary
}
