% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_functions.R
\name{summary.ssp.quantreg}
\alias{summary.ssp.quantreg}
\title{Main results summary}
\usage{
\method{summary}{ssp.quantreg}(object, ...)
}
\arguments{
\item{object}{A list object output by the main function, which contains the
results of the estimation of the parameters, the estimation of the
covariance matrix, subsample index, etc.}

\item{...}{Additional arguments passed to the summary function.}
}
\value{
A data frame will be printed.
}
\description{
Main results summary
}
\examples{
#logistic regression
set.seed(1)
}
