% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rare_events_logistic_functions.R
\name{summary.ssp.relogit}
\alias{summary.ssp.relogit}
\title{relogit Main results summary}
\usage{
\method{summary}{ssp.relogit}(object, ...)
}
\arguments{
\item{object}{A list object output by the main function, which contains the}

\item{...}{Additional arguments passed to the summary function.
results of the estimation of the parameters, the estimation of the
covariance matrix, subsample index, etc.}
}
\value{
A series of data.frame will be printed.
}
\description{
relogit Main results summary
}
\examples{
#TBD
}
