% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_main_function.R
\name{ssp.quantreg}
\alias{ssp.quantreg}
\title{Optimal Subsampling Methods for Quantile Regression Model}
\usage{
ssp.quantreg(
  formula,
  data,
  subset = NULL,
  tau = 0.5,
  n.plt,
  n.ssp,
  B = 5,
  boot = TRUE,
  criterion = "optL",
  sampling.method = "withReplacement",
  likelihood = c("weighted"),
  control = list(...),
  contrasts = NULL,
  ...
)
}
\arguments{
\item{formula}{An object of class "formula" which describes the model to be
fitted.}

\item{data}{A data frame containing the variables in the model.}

\item{subset}{An optional vector specifying a subset of observations to be used.}

\item{tau}{The quantile to be estimated,}

\item{n.plt}{The pilot subsample size (the first-step subsample size).
These samples will be used to estimate the pilot estimator as well as to
estimate the optimal subsampling probability.}

\item{n.ssp}{The expectation optimal subsample size (the second-step subsample
size). For \code{sampling.method = 'withReplacement'}, \code{n.ssp} is exactly the subsample size. For \code{sampling.method = 'poisson'}, \code{n.ssp} is the expectation of subsample size.}

\item{B}{The number of subsamples for the iterative sampling algorithm. Each subsample contains \code{n.ssp} observations. This allows us to estimate the covariance matrix.}

\item{boot}{If TRUE then perform iterative sampling algorithm and estimate the covariance matrix. If FALSE then only one subsample is returned which contains n.ssp observations.}

\item{criterion}{The criterion of optimal subsampling probabilities.
Choices include \code{optL}(default) and \code{uniform}.}

\item{sampling.method}{The sampling method for drawing the optimal subsample.
Choices include \code{withReplacement} and \code{poisson}(default).}

\item{likelihood}{The type of the maximum likelihood function used to
calculate the optimal subsampling estimator. Currently it uses \code{weighted}.}

\item{control}{A list of parameters for controlling the sampling process. Default is \code{list(alpha=0, b=2)}.}

\item{contrasts}{An optional list. It specifies how categorical variables are represented in the design matrix. For example, \code{contrasts = list(v1 = 'contr.treatment', v2 = 'contr.sum')}.}

\item{...}{A list of parameters which will be passed to \code{quantreg::rq()}.}
}
\value{
ssp.quantreg returns an object of class "ssp.quantreg" containing the following components (some are optional):

\describe{
\item{model.call}{model call}
\item{coef.plt}{pilot estimator}
\item{coef}{optimal subsample estimator.}
\item{cov}{covariance matrix of \code{coef}}
\item{index.plt}{index of pilot subsample in the full sample}
\item{index.ssp}{index of optimal subsample in the full sample}
\item{N}{number of observations in the full sample}
\item{subsample.size.expect}{expected subsample size}
\item{terms}{model terms}
}
}
\description{
Draw subsample from full dataset and fit quantile regression model on subsample.
}
\details{
Additional details... briefly introduce the idea.
}
\examples{
#quantile regression
set.seed(1)
N <- 1e4
B <- 5
tau <- 0.75
beta.true <- rep(1, 7)
d <- length(beta.true) - 1
corr  <- 0.5
sigmax  <- matrix(0, d, d)
for (i in 1:d) for (j in 1:d) sigmax[i, j] <- corr^(abs(i-j))
X <- MASS::mvrnorm(N, rep(0, d), sigmax)
err <- rnorm(N, 0, 1) - qnorm(tau)
Y <- beta.true[1] + X \%*\% beta.true[-1] + 
err * rowMeans(abs(X))
data <- as.data.frame(cbind(Y, X))
colnames(data) <- c("Y", paste("V", 1:ncol(X), sep=""))
formula <- Y ~ .
n.plt <- 200
n.ssp <- 100
optL.results <- ssp.quantreg(formula,data,tau = tau,n.plt = n.plt,
n.ssp = n.ssp,B = B,boot = TRUE,criterion = 'optL',
sampling.method = 'withReplacement',likelihood = 'weighted')
summary(optL.results)
uni.results <- ssp.quantreg(formula,data,tau = tau,n.plt = n.plt,
n.ssp = n.ssp,B = B,boot = TRUE,criterion = 'uniform',
sampling.method = 'withReplacement', likelihood = 'weighted')
summary(uni.results)
}
\references{
Wang, H., & Ma, Y. (2021). Optimal subsampling for quantile regression in big data. \emph{Biometrika}, \strong{108}(1), 99-112. \doi{https://doi.org/10.1093/biomet/asaa043}
}
