\name{subsampling-package}
\alias{subsampling-package}
\alias{subsampling}
\docType{package}
\title{
\packageTitle{subsampling}
}
\description{
\packageDescription{subsampling}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{subsampling}
\packageIndices{subsampling}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{subsampling}

Maintainer: \packageMaintainer{subsampling}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
