% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsampling_pkg.R
\docType{package}
\name{subsampling-package}
\alias{subsampling-package}
\alias{_PACKAGE}
\alias{Add}
\alias{later}
\title{Title: Add later}
\description{
More about what it does (maybe more than one line)
}
\references{
Add later
}
