% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsampling_pkg.R
\docType{package}
\name{subsampling}
\alias{subsampling}
\alias{subsampling-package}
\title{Title: Add later}
\description{
For common statistical models such as glm and quantile regression model, draw subsamples from full dataset by different optimal subsampling criterion and estimate coefficients and covariance matrix based on the subsample. When developing statistical models on massive datasets, subsampling is effective to reduce computational burden and optimal subsampling methods can improve estimation efficiency.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dqksnow/Subsampling}
  \item Report bugs at \url{https://github.com/dqksnow/Subsampling/issues}
}

}
\author{
\strong{Maintainer}: Qingkai Dong \email{qingkai.dong@uconn.edu}

Authors:
\itemize{
  \item Haiying Wang
  \item Yaqiong Yao
  \item Qiang Zhang
  \item Jun Yan
}

}
