% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsampling_pkg.R
\docType{package}
\name{subsampling}
\alias{subsampling}
\alias{_PACKAGE}
\alias{Add}
\alias{later}
\title{Title: Add later}
\description{
When developing statistical models on massive datasets, subsampling is effective to reduce computational burden and optimal subsampling methods can improve estimation efficiency. This package implements various optimal subsampling methods for statistical analysis, providing tools for efficient and accurate estimation in massive datasets.
}
\references{
Add later
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dqksnow/Subsampling}
  \item Report bugs at \url{https://github.com/dqksnow/Subsampling/issues}
}

}
\author{
\strong{Maintainer}: Qingkai Dong \email{qingkai.dong@uconn.edu}

Authors:
\itemize{
  \item Haiying Wang
  \item Yaqiong Yao
  \item Jun Yan
}

}
